feat: add settings to improve documentation archives generation (#42)
We don't want to display such links in documentation archives we want to make
available for download only.
In this case, the documentation content source code is not available locally
so keeping the links is just misleading.

For single component version archive, this is not need as we have a single
component. In addition, the link produced for component in this list doesn't
work for local file browsing.