feat: remove the 'isNotApiComponent' handlebars macro
We don't have component using the 'bonitaAPI' component.name. So, there is no
reason for filtering such component in nav and toc.
This seems to be a remaining part of the POC done in 2020: we were trying to
integrate the Bonita API documentation within the site. We don't have such API
in the site for now.