fix: don't underline link in breadcrumbs
In a previous change, we made the links in the toolbar message blocks more visible by underlining them. This introduced
a side effect on the links of the breadcrumbs that were also underlined.
This is now fixed and dedicated pages have been created to easily test the rendering of all kinds of messages.

In addition, refactor the CSS classes of the message blocks
- remove duplication
- apply underline to anchor inside the message blocks, not in the other toolbar elements
- introduce a dedicated css class for custom message block for clarity (even if the definition is the same as the one
for next version block)