ci: deploy to gh-pages with GitHub Actions
We don't need to keep a history of the built site, so we can get rid of the gh-pages branch.
In addition, we are sure to now have a fresh deploy. With the branch, some extra files are kept
except if they are removed manually.
Removing the branch will also reduce the size of the repository.