feat: provide link to related latest page for "out of support" pages
Previously, in the "out of support", the link to the "latest" resource was always targeting the home
page of the latest version of the component.
Now, the link targets the latest version of the current page when it exists.
The template use the "relativize" helper to make the url work with local file browsing

Also update the model-ui used for the preview to demonstrate the change.