feat: choose theme based on system preferences
Previously, there were only 2 states for the theme selector: dark or light.
If the user didn't touch the theme selector, the system preferences were used without they know it.
Once they use the theme selector, it was no longer possible to use the system preferences.

This is now possible to choose to use the theme based on the system preferences.
The ball selector has been replaced by an icon rolling, so only the current state of the theme selection is displayed.
The icons are taken from Material Symbols.